backend/requirements/common.txt
#django~=4.?
djangorestframework
pillow
django-cors-headers